Add github-pages gem
The deployment of the site using GitHub Pages is not rendered
correctly as evidenced by pages without titles appearing in the site
navigation and Liquid tags -- specifically output of variables --
appearing rather than the value of the variables. This behavior was
confirmed locally when building the site with the github-pages gem.

This change removes the github-pages gem from the `:jekyll_plugins`
group as suggested in an open issue.

References github/pages-gem#657