Disable jekyll-titles-from-headings gem
GitHub Pages enables the aforementioned plugin by default, but doing
so conflicts with the ability to create pages *without* a title and
have those pages omitted from the site's navigation (courtesy of
minima's `_includes/header.html`, which ignores pages that do not have
a title). Thus, this change disables the jekyll-titles-from-headings
plugin.