Add warning about unmaintained site
The websites for multiple course offerings are essentially identical,
which makes it difficult to know at a glance the corresponding
semester when viewing the site. This change comprises a banner at the
top of the site to indicate that the site is for a prior course
offering and is no longer maintained.