Add script to export Jekyll site
This change comprises a Bash script to automate the creation of a
local copy of a Jekyll site. Unfortunately, `jekyll build` uses
absolute URLs so `wget` transforms all the links to relative URLs,
which is required when viewing the site locally.

Because wget does not convert non-HTML links (such as JavaScript
redirects), the host (e.g., http://localhost:4000/) is removed from
URLs. This additional rewrite allows redirects to resolve properly
when viewing the site from the local file system. In addition, URLs
with a query string are *not* included.

Unmodified files are not overwritten in the destination, which allows
timestamps to be used to determine which files have been updated.